requirements.txt
gunicorn
openai>=1.0.0
Flask
python-dotenv
PyPDF2
weasyprint
pytest
pytest-mock
werkzeug
